refactor text area code editor into hook
